@@ -14,19 +14,19 @@
   "private": true,
   "dependencies": {
     "@adactive/bootstrap-tagsinput": "0.8.2",
-    "@angular/animations": "^13.2.6",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.2.6",
-    "@angular/common": "^13.2.6",
-    "@angular/compiler": "^13.2.6",
-    "@angular/core": "^13.2.6",
-    "@angular/elements": "^13.2.6",
-    "@angular/forms": "^13.2.6",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/elements": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/google-maps": "^13.3.3",
-    "@angular/localize": "^13.2.6",
+    "@angular/localize": "^14.3.0",
     "@angular/material": "^13.2.6",
-    "@angular/platform-browser": "^13.2.6",
-    "@angular/platform-browser-dynamic": "^13.2.6",
-    "@angular/router": "^13.2.6",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@ng-bootstrap/ng-bootstrap": "12.0.1",
     "@popperjs/core": "^2.11.4",
     "ajv": "7.0.1",
@@ -56,10 +56,10 @@
     "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.6",
-    "@angular/cli": "^13.2.6",
-    "@angular/compiler-cli": "^13.2.6",
-    "@angular/language-service": "13.2.6",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "14.3.0",
     "@types/bootstrap": "5.0.1",
     "@types/chartist": "0.11.0",
     "@types/crypto-js": "^4.1.1",
@@ -80,6 +80,6 @@
     "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "7.0.0",
     "ts-node": "~10.7.0",
-    "typescript": "~4.4.2"
+    "typescript": "~4.8.4"
   }
 }
